Backend completo

+++ backend/src/main/java/com/example/backend/Service/EstoqueService.java
@@ -1,8 +1,10 @@
 package com.example.backend.Service;
 
 import com.example.backend.Domain.Estoque;
+import com.example.backend.Exception.GenericHTTPException;
 import com.example.backend.Repository.EstoqueRepository;
 import org.springframework.stereotype.Service;
+import org.zalando.problem.Status;
 
 import java.util.List;
 
@@ -18,7 +20,11 @@ public class EstoqueService {
     }
 
     public List<Estoque> getAll(){
-        return estoqueRepository.findAll();
+        List<Estoque> produtos = estoqueRepository.findAll();
+        if(produtos == null || produtos.isEmpty()){
+            throw new GenericHTTPException("Nenhum produto encontrado", Status.NOT_FOUND);
+        }
+        return produtos;
     }
 
     public Estoque save(Estoque produto){
@@ -28,6 +34,9 @@ public class EstoqueService {
 
     public Estoque update(Long id, Estoque produto){
         var produtoVelho = estoqueRepository.findById(id);
+        if(produtoVelho == null){
+            throw new GenericHTTPException("Produto não encontrado", Status.NOT_FOUND);
+        }
         produtoVelho.setNome(produto.getNome());
         produtoVelho.setQuantidade(produto.getQuantidade());
         produtoVelho.setValor(produto.getValor());
@@ -36,6 +45,9 @@ public class EstoqueService {
 
     public void delete(Long id){
         var produtoDeletado = estoqueRepository.findById(id);
+        if(produtoDeletado == null){
+            throw new GenericHTTPException("Produto não encontrado", Status.NOT_FOUND);
+        }
         estoqueRepository.delete(produtoDeletado);
     }
 }
